few nits
