+++ praetorian_cli/sdk/model/utils.py
@@ -1,7 +1,5 @@
 # This file largely mirror the model definitions in the model package in chariot-client.
 
-import click    
-
 
 def asset_key(dns, name):
     return f'#asset#{dns}#{name}'
@@ -38,27 +36,3 @@ def configuration_key(name):
 
 def seed_status(type, status_code):
     return f'{type}#{status_code}'
-
-
-def get_dict_from_entries(entry):
-    config_dict = {}
-    for item in entry:
-        if '=' not in item:
-            click.echo(f"Error: Entry '{item}' is not in the format key=value")
-            return
-
-        if item.count('=') > 1:
-            click.echo(f"Error: Entry '{item}' contains multiple '=' characters. Format should be key=value")
-            return
-
-        key, value = item.split('=', 1)
-
-        if not key:
-            click.echo("Error: Key cannot be empty")
-            return
-
-        if not value:
-            click.echo("Error: Value cannot be empty")
-            return
-
-        config_dict[key] = value
